+++ native/src/consensus/agreement.rs
@@ -1,17 +1,17 @@
+use kodama::{linkage, Method};
 use std::collections::HashMap;
-use kodama::{Method, linkage};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Results {
     agreement_count: u32,
-    hit_count: u32
+    hit_count: u32,
 }
 
 impl Results {
     pub fn new() -> Self {
         Results {
             agreement_count: 0,
-            hit_count: 0
+            hit_count: 0,
         }
     }
 }
@@ -20,7 +20,7 @@ impl Results {
 pub struct Agreement {
     results: HashMap<String, Results>,
     threshold: u32,
-    sample_count: u32
+    sample_count: u32,
 }
 
 impl Agreement {
@@ -29,7 +29,7 @@ impl Agreement {
         Agreement {
             results,
             threshold,
-            sample_count: 0
+            sample_count: 0,
         }
     }
 
@@ -48,7 +48,10 @@ impl Agreement {
                 source_hit_count += 1;
                 labels.push(source);
                 coordinates.push(point);
-                self.results.entry(String::from(source)).or_insert(Results::new()).hit_count += 1;
+                self.results
+                    .entry(String::from(source))
+                    .or_insert(Results::new())
+                    .hit_count += 1;
             }
         }
 
@@ -65,7 +68,7 @@ impl Agreement {
                 condensed.push(haversine(*coordinates[row], *coordinates[col]));
             }
         }
-    
+
         // Perform hierarchical clustering and return the resulting dendrogram
         let dend = linkage(&mut condensed, coordinates.len(), Method::Single);
 
@@ -79,7 +82,7 @@ impl Agreement {
                 break;
             }
         }
-    
+
         // Re-associate indexes with source values, and update the agreement counts for each
         let modal_cluster: Vec<&String> = modal_cluster
             .into_iter()
@@ -87,7 +90,9 @@ impl Agreement {
             .map(|x| labels[x])
             .collect();
         for source in modal_cluster {
-            self.results.entry(String::from(source)).and_modify(|e| e.agreement_count += 1);
+            self.results
+                .entry(String::from(source))
+                .and_modify(|e| e.agreement_count += 1);
         }
     }
 }
@@ -101,8 +106,7 @@ fn haversine((lon1, lat1): (f64, f64), (lon2, lat2): (f64, f64)) -> f64 {
     let delta_lat = lat2 - lat1;
     let delta_lon = lon2 - lon1;
     let x =
-        (delta_lat / 2.0).sin().powi(2)
-        + lat1.cos() * lat2.cos() * (delta_lon / 2.0).sin().powi(2);
+        (delta_lat / 2.0).sin().powi(2) + lat1.cos() * lat2.cos() * (delta_lon / 2.0).sin().powi(2);
     2.0 * EARTH_RADIUS * x.sqrt().atan() * 1000.0
 }
 
@@ -115,15 +119,21 @@ mod tests {
         let mut agreement = Agreement::new(25);
 
         let mut source_map = HashMap::new();
-        source_map.insert(String::from("source1"), Some((-77.0013365,38.8959637)));
-        source_map.insert(String::from("source2"), Some((-77.0013338,38.8959407)));
-        source_map.insert(String::from("source3"), Some((-77.0013311,38.8955170)));
+        source_map.insert(String::from("source1"), Some((-77.0013365, 38.8959637)));
+        source_map.insert(String::from("source2"), Some((-77.0013338, 38.8959407)));
+        source_map.insert(String::from("source3"), Some((-77.0013311, 38.8955170)));
 
         agreement.process_points(&source_map);
 
-        source_map.entry(String::from("source1")).and_modify(|e| *e = Some((-77.0033025,38.8971410)));
-        source_map.entry(String::from("source2")).and_modify(|e| *e = Some((-77.0032677,38.8971390)));
-        source_map.entry(String::from("source3")).and_modify(|e| *e = Some((-77.0038872,38.8970513)));
+        source_map
+            .entry(String::from("source1"))
+            .and_modify(|e| *e = Some((-77.0033025, 38.8971410)));
+        source_map
+            .entry(String::from("source2"))
+            .and_modify(|e| *e = Some((-77.0032677, 38.8971390)));
+        source_map
+            .entry(String::from("source3"))
+            .and_modify(|e| *e = Some((-77.0038872, 38.8970513)));
 
         agreement.process_points(&source_map);
 
@@ -137,15 +147,21 @@ mod tests {
         let mut agreement = Agreement::new(25);
 
         let mut source_map = HashMap::new();
-        source_map.insert(String::from("source1"), Some((-76.9732081,38.9168672)));
-        source_map.insert(String::from("source2"), Some((-76.9733476,38.9163518)));
-        source_map.insert(String::from("source3"), Some((-76.9731089,38.9175434)));
+        source_map.insert(String::from("source1"), Some((-76.9732081, 38.9168672)));
+        source_map.insert(String::from("source2"), Some((-76.9733476, 38.9163518)));
+        source_map.insert(String::from("source3"), Some((-76.9731089, 38.9175434)));
 
         agreement.process_points(&source_map);
 
-        source_map.entry(String::from("source1")).and_modify(|e| *e = Some((-76.9717141,38.9309358)));
-        source_map.entry(String::from("source2")).and_modify(|e| *e = Some((-76.9710302,38.9312738)));
-        source_map.entry(String::from("source3")).and_modify(|e| *e = Some((-76.9720950,38.9308064)));
+        source_map
+            .entry(String::from("source1"))
+            .and_modify(|e| *e = Some((-76.9717141, 38.9309358)));
+        source_map
+            .entry(String::from("source2"))
+            .and_modify(|e| *e = Some((-76.9710302, 38.9312738)));
+        source_map
+            .entry(String::from("source3"))
+            .and_modify(|e| *e = Some((-76.9720950, 38.9308064)));
 
         agreement.process_points(&source_map);
 
@@ -159,15 +175,21 @@ mod tests {
         let mut agreement = Agreement::new(25);
 
         let mut source_map = HashMap::new();
-        source_map.insert(String::from("source1"), Some((-76.9732081,38.9168672)));
-        source_map.insert(String::from("source2"), Some((-76.9733476,38.9163518)));
+        source_map.insert(String::from("source1"), Some((-76.9732081, 38.9168672)));
+        source_map.insert(String::from("source2"), Some((-76.9733476, 38.9163518)));
         source_map.insert(String::from("source3"), None);
 
         agreement.process_points(&source_map);
 
-        source_map.entry(String::from("source1")).and_modify(|e| *e = None);
-        source_map.entry(String::from("source2")).and_modify(|e| *e = Some((-76.9710302,38.9312738)));
-        source_map.entry(String::from("source3")).and_modify(|e| *e = Some((-76.9720950,38.9308064)));
+        source_map
+            .entry(String::from("source1"))
+            .and_modify(|e| *e = None);
+        source_map
+            .entry(String::from("source2"))
+            .and_modify(|e| *e = Some((-76.9710302, 38.9312738)));
+        source_map
+            .entry(String::from("source3"))
+            .and_modify(|e| *e = Some((-76.9720950, 38.9308064)));
 
         agreement.process_points(&source_map);
 
@@ -191,16 +213,53 @@ mod tests {
         source_map.insert(String::from("Framingham"), Some((-71.4166667, 42.2791667)));
         source_map.insert(String::from("Marlborough"), Some((-71.5527778, 42.3458333)));
         source_map.insert(String::from("Northbridge"), Some((-71.6500000, 42.1513889)));
-        source_map.insert(String::from("Southborough"), Some((-71.5250000, 42.3055556)));
+        source_map.insert(
+            String::from("Southborough"),
+            Some((-71.5250000, 42.3055556)),
+        );
         source_map.insert(String::from("Westborough"), Some((-71.6166667, 42.2694444)));
 
         agreement.process_points(&source_map);
 
-        assert_eq!(agreement.results.get("Fitchburg").unwrap().agreement_count, 0);
-        assert_eq!(agreement.results.get("Framingham").unwrap().agreement_count, 1);
-        assert_eq!(agreement.results.get("Marlborough").unwrap().agreement_count, 1);
-        assert_eq!(agreement.results.get("Northbridge").unwrap().agreement_count, 1);
-        assert_eq!(agreement.results.get("Southborough").unwrap().agreement_count, 1);
-        assert_eq!(agreement.results.get("Westborough").unwrap().agreement_count, 1);
+        assert_eq!(
+            agreement.results.get("Fitchburg").unwrap().agreement_count,
+            0
+        );
+        assert_eq!(
+            agreement.results.get("Framingham").unwrap().agreement_count,
+            1
+        );
+        assert_eq!(
+            agreement
+                .results
+                .get("Marlborough")
+                .unwrap()
+                .agreement_count,
+            1
+        );
+        assert_eq!(
+            agreement
+                .results
+                .get("Northbridge")
+                .unwrap()
+                .agreement_count,
+            1
+        );
+        assert_eq!(
+            agreement
+                .results
+                .get("Southborough")
+                .unwrap()
+                .agreement_count,
+            1
+        );
+        assert_eq!(
+            agreement
+                .results
+                .get("Westborough")
+                .unwrap()
+                .agreement_count,
+            1
+        );
     }
 }