+++ native/src/stream/addr.rs
@@ -1,7 +1,7 @@
 use std::convert::From;
-use std::iter::Iterator;
-use std::io::{Write, BufWriter};
 use std::fs::File;
+use std::io::{BufWriter, Write};
+use std::iter::Iterator;
 
 use crate::{stream::geo::GeoStream, Address, Context};
 
@@ -9,7 +9,7 @@ pub struct AddrStream {
     context: Context,
     input: GeoStream,
     buffer: Option<Vec<u8>>, //Used by Read impl for storing partial features
-    errors: Option<BufWriter<File>>
+    errors: Option<BufWriter<File>>,
 }
 
 impl AddrStream {
@@ -20,8 +20,8 @@ impl AddrStream {
             buffer: None,
             errors: match errors {
                 None => None,
-                Some(path) => Some(BufWriter::new(File::create(path).unwrap()))
-            }
+                Some(path) => Some(BufWriter::new(File::create(path).unwrap())),
+            },
         }
     }
 }
@@ -38,7 +38,7 @@ impl std::io::Read for AddrStream {
             } else {
                 let feat = match self.next() {
                     Some(feat) => feat.to_tsv(),
-                    None => String::from("")
+                    None => String::from(""),
                 };
 
                 let mut bytes = feat.into_bytes();
@@ -83,9 +83,11 @@ impl Iterator for AddrStream {
 
                             Err(err)
                         }
-                    }
+                    },
                 },
-                None => { return None; }
+                None => {
+                    return None;
+                }
             };
         }