+++ native/src/stream/mod.rs
@@ -3,7 +3,7 @@ pub mod geo;
 pub mod net;
 pub mod poly;
 
-pub use self::poly::PolyStream;
-pub use self::geo::GeoStream;
 pub use self::addr::AddrStream;
+pub use self::geo::GeoStream;
 pub use self::net::NetStream;
+pub use self::poly::PolyStream;