+++ native/src/stream/poly.rs
@@ -1,14 +1,14 @@
 use std::convert::From;
-use std::iter::Iterator;
-use std::io::{Write, BufWriter};
 use std::fs::File;
+use std::io::{BufWriter, Write};
+use std::iter::Iterator;
 
 use crate::{stream::geo::GeoStream, Polygon};
 
 pub struct PolyStream {
     input: GeoStream,
     buffer: Option<Vec<u8>>, //Used by Read impl for storing partial features
-    errors: Option<BufWriter<File>>
+    errors: Option<BufWriter<File>>,
 }
 
 impl PolyStream {
@@ -18,8 +18,8 @@ impl PolyStream {
             buffer: None,
             errors: match errors {
                 None => None,
-                Some(path) => Some(BufWriter::new(File::create(path).unwrap()))
-            }
+                Some(path) => Some(BufWriter::new(File::create(path).unwrap())),
+            },
         }
     }
 }
@@ -36,7 +36,7 @@ impl std::io::Read for PolyStream {
             } else {
                 let feat = match self.next() {
                     Some(feat) => feat.to_tsv(),
-                    None => String::from("")
+                    None => String::from(""),
                 };
 
                 let mut bytes = feat.into_bytes();
@@ -81,9 +81,11 @@ impl Iterator for PolyStream {
 
                             Err(err)
                         }
-                    }
+                    },
                 },
-                None => { return None; }
+                None => {
+                    return None;
+                }
             };
         }