+++ native/src/text/replace.rs
@@ -1,6 +1,6 @@
-use fancy_regex::{Regex, Captures, Error};
-use std::str;
+use fancy_regex::{Captures, Error, Regex};
 use memchr::memchr;
+use std::str;
 
 pub trait ReplaceAll {
     fn replace_all(&self, text: &str, rep: &str) -> Result<String, Error>;
@@ -18,7 +18,7 @@ impl ReplaceAll for Regex {
                     None => {
                         new.push_str(&input);
                         break;
-                    },
+                    }
                     Some(m) => {
                         // capture group 0 always corresponds to the entire match
                         let pos = (m.pos(0).unwrap().0, m.pos(0).unwrap().1);
@@ -37,7 +37,7 @@ impl ReplaceAll for Regex {
                     None => {
                         new.push_str(&input);
                         break;
-                    },
+                    }
                     Some(m) => {
                         new.push_str(&input[..m.0]);
                         new.push_str(&rep);
@@ -63,11 +63,7 @@ struct CaptureRef {
     end: usize,
 }
 
-fn expand_str(
-    caps: &Captures,
-    mut replacement: &str,
-    dst: &mut String,
-) {
+fn expand_str(caps: &Captures, mut replacement: &str, dst: &mut String) {
     while !replacement.is_empty() {
         match memchr(b'$', replacement.as_bytes()) {
             None => break,
@@ -99,9 +95,7 @@ fn expand_str(
 /// starting at the beginning of `replacement`.
 ///
 /// If no such valid reference could be found, None is returned.
-fn find_cap_ref<T: ?Sized + AsRef<[u8]>>(
-    replacement: &T,
-) -> Option<CaptureRef> {
+fn find_cap_ref<T: ?Sized + AsRef<[u8]>>(replacement: &T) -> Option<CaptureRef> {
     let mut i = 0;
     let rep: &[u8] = replacement.as_ref();
     if rep.len() <= 1 || rep[0] != b'$' {
@@ -118,17 +112,14 @@ fn find_cap_ref<T: ?Sized + AsRef<[u8]>>(
     // We just verified that the range 0..cap_end is valid ASCII, so it must
     // therefore be valid UTF-8. If we really cared, we could avoid this UTF-8
     // check with either unsafe or by parsing the number straight from &[u8].
-    let cap = str::from_utf8(&rep[i..cap_end])
-                  .expect("valid UTF-8 capture name");
+    let cap = str::from_utf8(&rep[i..cap_end]).expect("valid UTF-8 capture name");
 
     match cap.parse::<u32>() {
-        Ok(i) => {
-            Some(CaptureRef {
-                cap: i as usize,
-                end: cap_end
-            })
-        },
-        Err(_) => None
+        Ok(i) => Some(CaptureRef {
+            cap: i as usize,
+            end: cap_end,
+        }),
+        Err(_) => None,
     }
 }
 
@@ -136,7 +127,7 @@ fn find_cap_ref<T: ?Sized + AsRef<[u8]>>(
 /// Modified to only support numbered capture groups
 fn is_valid_cap(b: &u8) -> bool {
     match *b {
-        b'0' ..= b'9' => true,
+        b'0'..=b'9' => true,
         _ => false,
     }
 }
@@ -147,17 +138,71 @@ mod tests {
 
     #[test]
     fn test_replace() {
-        assert_eq!(Regex::new(r"(?:floor|fl) #?\d{1,3}").unwrap().replace_all("123 main st", "").unwrap(), "123 main st");
+        assert_eq!(
+            Regex::new(r"(?:floor|fl) #?\d{1,3}")
+                .unwrap()
+                .replace_all("123 main st", "")
+                .unwrap(),
+            "123 main st"
+        );
         assert_eq!(Regex::new(r"(?:apartment|apt|bldg|building|rm|room|unit) #?(?:[A-Z]|\d+|[A-Z]\d+|\d+[A-Z]|\d+-\d+[A-Z]?)").unwrap().replace_all("123 main st apt #5 washington dc", "").unwrap(), "123 main st  washington dc");
-        assert_eq!(Regex::new(r"(?:floor|fl) #?\d{1,3}").unwrap().replace_all("123 main st floor 5", "").unwrap(), "123 main st ");
-        assert_eq!(Regex::new(r"\d{1,3}(?:st|nd|rd|th) (?:floor|fl)").unwrap().replace_all("5th floor", "").unwrap(), "");
-        assert_eq!(Regex::new(r"[１1]丁目").unwrap().replace_all("1丁目 意思", "一丁目").unwrap(), "一丁目 意思");
+        assert_eq!(
+            Regex::new(r"(?:floor|fl) #?\d{1,3}")
+                .unwrap()
+                .replace_all("123 main st floor 5", "")
+                .unwrap(),
+            "123 main st "
+        );
+        assert_eq!(
+            Regex::new(r"\d{1,3}(?:st|nd|rd|th) (?:floor|fl)")
+                .unwrap()
+                .replace_all("5th floor", "")
+                .unwrap(),
+            ""
+        );
+        assert_eq!(
+            Regex::new(r"[１1]丁目")
+                .unwrap()
+                .replace_all("1丁目 意思", "一丁目")
+                .unwrap(),
+            "一丁目 意思"
+        );
 
-        assert_eq!(Regex::new(r"([a-z]+)vagen").unwrap().replace_all("123 main st", "$1v").unwrap(), "123 main st");
-        assert_eq!(Regex::new(r"([a-z]+)vagen").unwrap().replace_all("amanuensvagen 5 stockholm sweden", "$1v").unwrap(), "amanuensv 5 stockholm sweden");
-        assert_eq!(Regex::new(r"([a-z]+)vagen").unwrap().replace_all("amanuensvagen 5 stockholm sweden gutenvagen", "$1v").unwrap(), "amanuensv 5 stockholm sweden gutenv");
-        assert_eq!(Regex::new(r"([^ ]+)(strasse|straße|str)").unwrap().replace_all("wilhelmstraße 3", "$1 str").unwrap(), "wilhelm str 3");
-        assert_eq!(Regex::new(r"(foo) (bar)").unwrap().replace_all("foo bar", "$2 $1").unwrap(), "bar foo");
+        assert_eq!(
+            Regex::new(r"([a-z]+)vagen")
+                .unwrap()
+                .replace_all("123 main st", "$1v")
+                .unwrap(),
+            "123 main st"
+        );
+        assert_eq!(
+            Regex::new(r"([a-z]+)vagen")
+                .unwrap()
+                .replace_all("amanuensvagen 5 stockholm sweden", "$1v")
+                .unwrap(),
+            "amanuensv 5 stockholm sweden"
+        );
+        assert_eq!(
+            Regex::new(r"([a-z]+)vagen")
+                .unwrap()
+                .replace_all("amanuensvagen 5 stockholm sweden gutenvagen", "$1v")
+                .unwrap(),
+            "amanuensv 5 stockholm sweden gutenv"
+        );
+        assert_eq!(
+            Regex::new(r"([^ ]+)(strasse|straße|str)")
+                .unwrap()
+                .replace_all("wilhelmstraße 3", "$1 str")
+                .unwrap(),
+            "wilhelm str 3"
+        );
+        assert_eq!(
+            Regex::new(r"(foo) (bar)")
+                .unwrap()
+                .replace_all("foo bar", "$2 $1")
+                .unwrap(),
+            "bar foo"
+        );
     }
 
     /// Tests from the core Rust regex crate
@@ -178,7 +223,10 @@ mod tests {
 
     macro_rules! c {
         ($name_or_number:expr, $pos:expr) => {
-            CaptureRef { cap: $name_or_number, end: $pos }
+            CaptureRef {
+                cap: $name_or_number,
+                end: $pos,
+            }
         };
     }